Application/Backend/_AuthService/requirements.txt
cffi==1.10.0
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyasn1==0.3.3
pycparser==2.18
PyJWT==1.5.2
six==1.10.0
Werkzeug==0.12.2
